### 2018-06-08 #224 @cboettig
submission: arkdb
Read in large, compressed text files into a databases in chunks.  Write all tables in a database out into compressed tsv files.

```
Package: arkdb
Version: 0.0.0.9000
Title: Archive and Unarchive Databases Using Flat Files
Description: Flat text files provide a more robust, compressable,
  and portable way to store tables.  This package provides convenient
  functions for exporting tables from relational database connections
  into compressed text files and streaming those text files back into
  a database without requiring the whole table to fit in working memory.
Authors@R: c(person("Carl", "Boettiger",
                    email = "cboettig@gmail.com",
                    comment=c(ORCID = "0000-0002-1642-628X")),
              person("Richard", "FitzJohn", role = "ctb"))
RoxygenNote: 6.0.1.9000
    dbplyr,
    DBI,
    RSQLite,
    progress,
    methods
    nycflights13,
    covr,
    fs,
```

https://github.com/cboettig/arkdb

data deposition! Because good archiving formats are crucial.  Kinda crazy how many researchers and repositories  distribute data only through APIs and uncompressed sql dumps.

Researchers working with large data files

I don't think so?

```

R CMD check results
0 errors | 0 warnings | 0 notes

```

  - @richfitz (I've listed him as a contributor because I've borrowed code from his other packages here, so I don't think that counts as a conflict of interest(?))
  - @jimhester
  - Maybe an easy package for a first-time reviewer, since this is rather minimal?
