### 2018-06-10 04814c2
support all 4 standard compression formats
