+++ vignettes/arkdb_intro.Rmd
@@ -26,7 +26,7 @@ Increasing data sizes create challenges for the fundamental tasks of publishing,
 Working with tables that are too large for working memory on most machines by using external relational database stores is now a common R practice, thanks to ever-rising availability of data and increasing support and popularity of packages such as `DBI`, `dplyr`, and `dbplyr`.  Working with plain text files becomes increasingly difficult in this context.  Many R users will not have sufficient RAM to simply read in a 10 GB `.tsv` file into R.  Similarly, moving a 10 GB database out of a relational data file and into a plain text file for archiving and distribution is similarly challenging from R. While most relational database back-ends implement some form of `COPY` or `IMPORT` that allows them to read in and export out plain text files directly, these methods are not consistent across database types and not part of the standard SQL interface.  Most importantly for our case, they also cannot be called directly from R, but require a separate stand-alone installation of the database client.  `arkdb` provides a simple solution to these two tasks. 
  
 
-The goal of `arkdb` is to provide a convenient way to move data from large compressed text files (`.tsv.bz2`) into any DBI-compliant database connection (see [DBI](https://db.rstudio.com/dbi/)), and move tables out of such databases into text files. The key feature of `arkdb` is that files are moved between databases and text files in chunks of a fixed size, allowing the package functions to work with tables that would be much to large to read into memory all at once.  This will be slower than reading the file into memory at one go, but can be scaled to larger data and larger data with no additional memory requirement. 
+The goal of `arkdb` is to provide a convenient way to move data from large compressed text files (e.g. `*.tsv.bz2`) into any DBI-compliant database connection (see [DBI](https://db.rstudio.com/dbi/)), and move tables out of such databases into text files. The key feature of `arkdb` is that files are moved between databases and text files in chunks of a fixed size, allowing the package functions to work with tables that would be much to large to read into memory all at once.  This will be slower than reading the file into memory at one go, but can be scaled to larger data and larger data with no additional memory requirement. 
 
 
 ## Installation
@@ -103,7 +103,7 @@ To import a file into an existing database, use `unark_file` instead, which take
 
 ## A note on compression
 
-`arkdb` uses the `bz2` compression algorithm, supported in base R, to compress `tsv` files.  `bz2` offers excellent compression levels, and while it is considerably slower to compress than other popular formats like `gzip`, it is comparably fast to uncompress.  Future versions of `arkdb` may provide more flexibility regarding compression. 
+`unark` can read from a variety of compression formats recognized by base R: `bzip2`, `gzip`, `zip`, and `xz`, and `ark` can choose any of these as the compression algorithm.  Note that there is some trade-off between speed of compression and efficiency (i.e. the final file size).  `ark` uses the `bz2` compression algorithm by default, supported in base R, to compress `tsv` files.  The  `bz2` offers excellent compression levels, but is considerably slower to compress than `gzip` or `zip`.  It is comparably fast to uncompress.  For faster archiving when maximum file size reduction is not critical, `gzip` will give nearly as effective compression in significantly less time.
 
 
 ## Distributing data
