+++ NAMESPACE
@@ -19,3 +19,4 @@ importFrom(dplyr,tbl)
 importFrom(methods,is)
 importFrom(progress,progress_bar)
 importFrom(readr,read_tsv)
+importFrom(tools,file_ext)
