### 2018-06-10 ec3c10a
unark takes a db connection directly
This more nicely parallels the design of ark, and makes unark_file unnecessary.  Also avoids any default to SQLite, the user can provide whatever database they want.
