+++ README.Rmd
@@ -60,7 +60,6 @@ Create an archive of the database:
 
 ```{r}
 ark(db, ".", lines = 50000)
-
 ```
 
 ## Unarchive
@@ -70,7 +69,9 @@ Import a list of compressed tabular files (i.e. `*.tsv.bz2`) into a local SQLite
 
 ```{r}
 files <- fs::dir_ls(glob = "*.tsv.bz2")
-new_db <-  unark(files, dbname = "local.sqlite", lines = 50000)
+new_db <- src_sqlite("local.sqlite")
+
+unark(files, new_db, lines = 50000)
 
 new_db
 ```
