+++ DESCRIPTION
@@ -24,7 +24,6 @@ Imports:
     dbplyr,
     DBI,
     readr,
-    RSQLite,
     progress,
     dplyr,
     methods,
