+++ DESCRIPTION
@@ -34,4 +34,5 @@ Suggests:
     knitr,
     covr,
     fs,
-    rmarkdown
+    rmarkdown,
+    RSQLite
