### 2018-06-11 63a62a2
minor tweaks
- consistent argument order in unark
- detect db type for windowing
- allow passing options to read_tsv (in a pinch)
