### 2018-06-11 #224 @lmullen
### Editor checks:

---

#### Editor comments

Hi @cboettig. Thanks for the submission. I'll be editing this package.

I get errors in the tests or examples when running `devtools::check(cran = FALSE, run_dont_test = TRUE)`. My guess is that those errors are in the `\dontrun{}` sections of the examples. Could you please fix those errors before the review starts?

One issue raised by `goodpractice::g()`. Could you look into this please?

```
fix this R CMD check NOTE: Namespace in Imports field not imported from:
    ‘dbplyr’ All declared Imports should be used.
```

I also note some genuine spelling errors. You may wish to look at `devtools::spell_check()` yourself.

```
avialable      ark.Rd:29
databaase      unark.Rd:5,89
ouput          ark.Rd:12
```

I will be looking for reviewers.
