### 2018-06-11 1a33432
Fix issues highlighted by @lmullen
- donttest examples should be happy now
- spelling fixed
- suggest, not import, dbplyr
