+++ R/ark.R
@@ -2,7 +2,7 @@
 #' Archive tables from a database as flat files
 #' 
 #' @param db_con a database connection
-#' @param dir a directory where we will write the compressed text files ouput
+#' @param dir a directory where we will write the compressed text files output
 #' @param lines the number of lines to use in each single chunk
 #' @param compress file compression algorithm. Should be one of "bzip2" (default),
 #' "gzip" (faster write times, a bit less compression), or "xz".
@@ -13,7 +13,7 @@
 #' is probably why you are using a database in the first place!)  Compressed
 #' text files will likely take up much less space, making them easier to store and
 #' transfer over networks.  Compressed plain-text files are also more archival
-#' friendly, as they rely on widely avialable and long-established open source compression
+#' friendly, as they rely on widely available and long-established open source compression
 #' algorithms and plain text, making them less vulnerable to loss by changes in 
 #' database technology and formats. 
 #' 
@@ -23,17 +23,15 @@
 #' 
 #' @examples 
 #' \donttest{
-#' sql_path <- tempdir()
-#' db <- dbplyr::nycflights13_sqlite(sql_path)
-#' dir <- file.path(tempdir(), "nycflights")
-#' dir.create(dir)
+#' # setup
+#' library(dplyr)
+#' dir <- tempdir() 
+#' db <- dbplyr::nycflights13_sqlite(tempdir())
 #' 
+#' ## And here we go:
 #' ark(db, dir)
-#' list.files(dir)
 #' 
 #' 
-#' ## clean up
-#' unlink(dir, TRUE)
 #' }
 ark <- function(db_con, dir, lines = 10000L, 
                 compress = c("bzip2", "gzip", "xz")){
