+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: arkdb
 Version: 0.0.0.9000
 Title: Archive and Unarchive Databases Using Flat Files
-Description: Flat text files provide a more robust, compressable,
+Description: Flat text files provide a more robust, compressible,
   and portable way to store tables.  This package provides convenient
   functions for exporting tables from relational database connections
   into compressed text files and streaming those text files back into
@@ -21,7 +21,6 @@ VignetteBuilder: knitr
 RoxygenNote: 6.0.1.9000
 Roxygen: list(markdown = TRUE)
 Imports: 
-    dbplyr,
     DBI,
     readr,
     progress,
@@ -29,6 +28,7 @@ Imports:
     methods,
     tools
 Suggests: 
+    dbplyr,
     nycflights13,
     testthat,
     knitr,
