### 2018-06-11 #224 @cboettig
@lmullen I continue to learn tricks from you!  should remember to run the `donttest`s and spell checks.  I think all these issues should be fixed in the current `master`, https://github.com/cboettig/arkdb/commit/1a33432c71befda38364be05163e7ad275d11d94
