### 2018-06-11 #224 @lmullen
@cboettig Thanks for fixing those. Everything looks good to start the review, and I've solicited reviewers.
