### 2018-06-13 35b853c
mysql doesn't have windowing either also provide generic escape to disable windowing
