### 2018-06-13 7c5e016
allow a list of tables instead of forcing whole db
