### 2018-06-15 #224 @lmullen
Reviewers: @richfitz and @expectopatronum.
Due date: 2018-07-06

Thanks to Verena and Rich for being willing to do this review. I've set a due date 3 weeks from now. Here are the [reviewer's guide](https://ropensci.github.io/dev_guide/onboarding-guide-for-reviewers.html) and the [reviewer's template](https://ropensci.github.io/dev_guide/reviewtemplate.html).

@cboettig Could you please add the rOpenSci review badge to your repository if you haven't done so already?
