+++ README.md
@@ -6,6 +6,7 @@ status](https://codecov.io/gh/cboettig/arkdb/branch/master/graph/badge.svg)](htt
 [![AppVeyor Build
 Status](https://ci.appveyor.com/api/projects/status/github/cboettig/arkdb?branch=master&svg=true)](https://ci.appveyor.com/project/cboettig/arkdb)
 [![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/arkdb)](https://cran.r-project.org/package=arkdb)
+[![](https://badges.ropensci.org/224_status.svg)](https://github.com/ropensci/onboarding/issues/224)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
@@ -55,11 +56,6 @@ Consider the `nycflights` database in SQLite:
 ``` r
 db <- dbplyr::nycflights13_sqlite(".")
 #> Caching nycflights db at ./nycflights13.sqlite
-#> Creating table: airlines
-#> Creating table: airports
-#> Creating table: flights
-#> Creating table: planes
-#> Creating table: weather
 ```
 
 Create an archive of the database:
@@ -68,44 +64,46 @@ Create an archive of the database:
 ark(db, ".", lines = 50000)
 #> Importing in 50000 line chunks:
 #> airlines
-#> ...Done! (in 0.011554 secs)
+#> ...Done! (in 0.01724696 secs)
 #> Importing in 50000 line chunks:
 #> airports
-#> ...Done! (in 0.05471492 secs)
+#> ...Done! (in 0.03898597 secs)
 #> Importing in 50000 line chunks:
 #> flights
-#> ...Done! (in 0.5434999 secs)
+#> ...Done! (in 0.553968 secs)
 #> Importing in 50000 line chunks:
 #> planes
-#> ...Done! (in 0.101465 secs)
+#> ...Done! (in 0.09305 secs)
 #> Importing in 50000 line chunks:
 #> weather
-#> ...Done! (in 0.321594 secs)
+#> ...Done! (in 0.2292418 secs)
 ```
 
 ## Unarchive
 
-Import a list of compressed tabular files (`*.tsv.bz2`) into a local
-SQLite database:
+Import a list of compressed tabular files (i.e. `*.tsv.bz2`) into a
+local SQLite database:
 
 ``` r
 files <- fs::dir_ls(glob = "*.tsv.bz2")
-new_db <-  unark(files, dbname = "local.sqlite", lines = 50000)
+new_db <- src_sqlite("local.sqlite", create=TRUE)
+
+unark(files, new_db, lines = 50000)
 #> Importing in 50000 line chunks:
 #> airlines.tsv.bz2
-#> ...Done! (in 0.05960894 secs)
+#> ...Done! (in 0.06284809 secs)
 #> Importing in 50000 line chunks:
 #> airports.tsv.bz2
-#> ...Done! (in 0.0306232 secs)
+#> ...Done! (in 0.0897131 secs)
 #> Importing in 50000 line chunks:
 #> flights.tsv.bz2
-#> ...Done! (in 0.2614648 secs)
+#> ...Done! (in 0.134465 secs)
 #> Importing in 50000 line chunks:
 #> planes.tsv.bz2
-#> ...Done! (in 0.1022561 secs)
+#> ...Done! (in 0.055408 secs)
 #> Importing in 50000 line chunks:
 #> weather.tsv.bz2
-#> ...Done! (in 0.173305 secs)
+#> ...Done! (in 0.168793 secs)
 
 new_db
 #> src:  sqlite 3.22.0 [local.sqlite]
