### 2018-06-19 24f9f12
support uncompressed as an option closes #2
