### 2018-06-19 da4b019
update docs re no compression option
