### 2018-06-27 ropensci/arkdb#3 @cboettig
Support more additional formats (i.e. fst, feather, etc?)
- [x] arkdb currently hardwires `write_tsv`/`read_tsv` for I/O.  This should be pluggable.

*done, now supports readr and base utils table I/O*

Would be nice to take advantage of things like faster speed of `fst::read_fst` /`fst::write_fst`, but looks like `fst` does not support the ability to append, making it impossible to stream in chunks.
