+++ R/unark.R
@@ -12,7 +12,6 @@
 #' @param db_con a database src (`src_dbi` object from `dplyr`)
 #' @param lines number of lines to read in a chunk.
 #' @inheritDotParams readr::read_tsv
-#' @importFrom DBI dbConnect
 #' @details `unark` will read in a files in chunks and 
 #' write them into a database.  This is essential for processing
 #' large compressed tables which may be too large to read into
