### 2018-07-07 #224 @cboettig
@expectopatronum Thanks for this awesome review :tada:.  Very helpful.  I thought I'd wait for @richfitz comments but might as well check these off now and asking a few follow-up questions.  (Rich has given me a hint meanwhile on at least one other thing I should improve).

> - no documentation/help for ?arkdb is provided

Good catch, fixed!

> - consider adding tests for error cases

I'm terrible at adding these, at least until I get bug reports on particular cases.  Are there some obvious example scenarios you can think of that should provide better error handling or error messages?

> - `@return` is not documented for ark
> The order of parameters in ark(db_con, dir, lines) and unark(files, db_con, lines) seems not intuitive for me. Unless it makes more sense for using piping (magrittr) with db_con %>% ark and files %>% unark, I would change the order to (db_con, dir/files, lines)

Grouping these, since they are related (`ark` returns a directory/path, now documented!, `unark` returns the `db_con` -- I think there's a good reason for this but I see the confusion too. As you noted, I expect most uses not to actually use the return object, so it is returned invisibly, as something that could potentially be used in piping.  For instance, imagine using arkdb to migrate from, say, a MySQL DB (ick!) to an MonetDBLite database:

```r
old <-  dbConnect(MySQL(), ...)
new <-  dbConnect(MonetDBLite(), "my_db")

old %>%
  ark( tmpdir() ) %>%
  fs::dir_ls() %>%
  unark(new)
```

This is why `ark` returns `dir` while `unark` returns `db_con`, and is also why `ark` takes the arguments in the order `db_con`, `dir`, while `unark` takes `files` first and `db_con` as the second argument.  I could be talked out of this, as I can see how reversing the order is confusing, but I think it as:  the first argument is the *input* (as required by piping), and the *second* is the destination.  `ark` operates on an existing database and outputs files, `unark` operates on files.  What do you think?

> `repeat` vs `while`  Yeah, looks like I could have used a `while` loop -- but as this is the block of code I stole from @richfitz , I'll let him comment as there may be a good case on `repeat`.  I'm not well-informed on the difference here.

Okay, pushing these few edits, lemme know what you think!
