### 2018-07-07 6dbfdce
Onboarding edits suggested by @expectopatronum
- Add pkg-level docs
- fix missing return on ark
- fix vignette name (I think)?
