### 2018-07-07 047da1f
handle DBI or dplyr connections in unark
