+++ NAMESPACE
@@ -2,15 +2,15 @@
 
 export(ark)
 export(unark)
+importFrom(DBI,dbFetch)
 importFrom(DBI,dbListTables)
+importFrom(DBI,dbSendQuery)
 importFrom(DBI,dbWriteTable)
 importFrom(dplyr,between)
 importFrom(dplyr,collect)
 importFrom(dplyr,filter)
-importFrom(dplyr,n)
 importFrom(dplyr,row_number)
 importFrom(dplyr,sql)
-importFrom(dplyr,summarise)
 importFrom(dplyr,tbl)
 importFrom(methods,is)
 importFrom(progress,progress_bar)
