### 2018-07-07 c7c0b81
fix BUG in line counting
