### 2018-07-07 b795b17
Drop hard dependency on dplyr
