+++ tests/testthat/test-arkdb.R
@@ -13,19 +13,23 @@ testthat::test_that("we can ark and unark a db", {
   files <- fs::dir_ls(dir, glob = "*.tsv.bz2")
   testthat::expect_length(files, 5)
 
+  myflights <- suppressMessages(
+    readr::read_tsv(fs::path(dir, "flights.tsv.bz2")))
+  testthat::expect_equal(dim(myflights), 
+                         dim(nycflights13::flights))
+  
   ## unark
   new_db <- src_sqlite("local.sqlite", create = TRUE)
   unark(files, new_db, lines = 50000)
   
   myflights <- tbl(new_db, "flights")
+  testthat::expect_is(myflights, "tbl_dbi")
   
+  myflights <- collect(myflights)
   testthat::expect_equal(dim(myflights), 
                          dim(nycflights13::flights))
   
  
-  testthat::expect_is(myflights, "tbl_dbi")
-
-  myflights <- collect(myflights)
   dim(myflights[[1]])
   ## Classes not preserved, we get read_tsv guesses on class
   
