+++ NAMESPACE
@@ -6,12 +6,6 @@ importFrom(DBI,dbFetch)
 importFrom(DBI,dbListTables)
 importFrom(DBI,dbSendQuery)
 importFrom(DBI,dbWriteTable)
-importFrom(dplyr,between)
-importFrom(dplyr,collect)
-importFrom(dplyr,filter)
-importFrom(dplyr,row_number)
-importFrom(dplyr,sql)
-importFrom(dplyr,tbl)
 importFrom(methods,is)
 importFrom(progress,progress_bar)
 importFrom(readr,read_tsv)
