### 2018-07-16 #224 @richfitz
**RGF**: As @cboettig notes above, this is adapted from a similar bit of code I wrote a few months back but I don't think that counts as a COI as I regularly hate on my own code...

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

- [ ] ~**Performance:** Any performance claims of the software been confirmed.~

---

I have not read the other review as I write this review (and having now looked though the other review it looks like Carl already mostly dealt with those comments).  Apologies this is so late

This package provides an interface for loading data into a database from disk, and extracting data from a database to disk, via R, without needing to do the entire operation in memory.  This will help with very large databases (on the order of hundreds of millions of rows or more) where fitting data into memory is not an option.  It will also increase the responsiveness of such operations (probably at the cost of overal speed) making large scale import and extraction from R a more pleasant experience.

This package, with about a hundred lines of code and two exported functions, fits a "micropackage" description, which makes reviewing a lot easier.

#### Major comments

The biggest flaw I see with the package is that it violates the [Open–closed principle](https://en.wikipedia.org/wiki/Open%E2%80%93closed_principle) (I know this is not OOP but I feel this is one that really transcends any paradigm).  Carl has used `readr::write_tsv` and `readr::read_tsv` for the archive/import - presumably they satisfy his personal preferences for speed/reliability/future compatibility.  But there's no real reason why `utils::read.csv`/`utils::write.csv` might not be used, or `datatable::fread` or even something like feather probably (I don't really see that `readr` is a necessary strong dependency here either).

The package could be extended to take either 3 arguments (or some sort of wrapper object) that includes an extension and two functions (read/write).  This interacts somewhat with the compression handling but I think it's probably still doable.  As it is, if someone has an aversion to `tsv` (e.g., they want to point something that requires `csv` at the archived database) they cannot use the package without changing the source.

#### Misc comments:

The example in the `README` produces a significant number of warnings: `Warning: Closing open result set, pending rows` - these should be silenced if expected, eliminated if not.

The progress and general chattyness needs to be tuneable (e.g., with a `verbose` and a `progress` argument.  Ideally both would be tuneable as progress bars are not very log-friendly but messages are (you can see this play out a bit with the vignette build - the progress output ends up on the parent terminal for me)

The log output as it is not incredibly informative after it has been run:

```
Exporting in 50000 line chunks:
airlines
...Done! (in 0.005350828 secs)
Exporting in 50000 line chunks:
airports
...Done! (in 0.02871299 secs)
Exporting in 50000 line chunks:
flights
[\] chunk 7...Done! (in 5.914837 secs)
Exporting in 50000 line chunks:
planes
...Done! (in 0.05566597 secs)
Exporting in 50000 line chunks:
weather
...Done! (in 0.4096012 secs)
```

Consider a little bit of reordering and indenting to get to

```
Exporting 'artlines' in 50000 line chunks:
   ...Done! (in 0.005350828 secs)
Exporting 'airports' in 50000 line chunks:
   ...Done! (in 0.02871299 secs)
Exporting 'flights' in 50000 line chunks:
   [\] chunk 7...Done! (in 5.914837 secs)
```

(and so on)

How is the default of `lines` chosen?  Is there some heuristic for best performance?  My gut instinct is that as it increases in size, the performance will get better, to a point.  10K lines might well be at the point where increasing the chunk size has limited benefit but I am curious if you looked?  (I think I settled around a similar number - perhaps 50K in my application).

In `ark.R`

``` r
  if (is(db_con, "SQLiteConnection") |
      is(db_con, "MySQLConnection") |
      Sys.getenv("arkdb_windowing") == "FALSE") {
    query <- paste("SELECT * FROM", tablename, "LIMIT",
                   lines, "OFFSET", (start-1)*lines)
  ## Postgres can do windowing
    query <- paste("SELECT * FROM",
                   tablename,
                    "WHERE rownum BETWEEN",
                   (start - 1) * lines,
                   "AND",
                   start * lines)
  }
```

(The `|` should be `||`).  But this also surely is not a complete test?  There will be other DBI wrappers (thinking here of MariaDB, there are at least two Postgres DB wrappers - so there may be similar duplication among MySQL).  I think that the polarity of this test is also incorrect.  Every SQL dialect can do the first form so that should be the fallback, not the Postgres extension.  So rather than testing "known not to support" test for "known to support".  You might also include a pacakge-cached capability test (keyed against class) for a `SELECT * FROM table WHERE rownum BETWEEN 1 and 2` - if you don't get a syntax error it should be ok to use.

#### Super minor comments

Ij `unark.R:normalize_con`

``` r
if(is(db_con, "src_dbi")){
```

I though that `inherits` was still the way to test this?

Your postgres windowed code is not tested.  You might try that on travis?

Your `normalize_con` code is not tested for the non-`src_dbi` path

I am confused by the point of the untested early exit in `unark_file` - when would this be wanted?

There are also a few error paths untested (see covr output)

In `unark.R`

```r
    d <- reader()
    body <- paste0(c(header, d$data), "\n", collapse = "")
    p$tick()
    chunk <- readr::read_tsv(body, ...)
    DBI::dbWriteTable(db_con, tbl_name, chunk, append=TRUE)
```

I might be missing something but surely one can compute the headers ahead of time (0'th iteration) and then do:

``` r
    chunk <- reader()
    names(chunk) <- header
    p$tick()
    DBI::dbWriteTable(db_con, tbl_name, chunk, append=TRUE)
```

with modifications to `read_chunked` like

```r
read_chunked <- function(con, n) {
  assert_connection(con)
  next_chunk <- readLines(con, n)
  if (length(next_chunk) == 0L) {
    # if we don't stop, we will hit an error!
    stop("connection has already been completely read")
  }
  function() {
    data <- next_chunk
    next_chunk <<- read_tsv(con, n, header = FALSE) # <-- change here
    complete <- length(next_chunk) == 0L # <-- probable change here too
    list(data = data, complete = complete)
  }
}
```

(This is quite possibly something I could do in my code that this comes from too!)

I would be probably inclined to split `ark_chunk` into a "read" and "save" step.
