### 2018-07-16 #224 @cboettig
@richfitz Thanks, this is excellent and the detail much appreciated!  Addresses a bunch of things that have puzzled or bugged me so really appreciate your expertise here.

 Just skimming review for now, but one point where I could use a bit of additional feedback is the hard-wriring `read_tsv()` bit.  I agree that should be abstracted out (it's [issue #3](https://github.com/cboettig/arkdb/issues/3))), but could use a bit more guidance/ideas on how to make it more general, how general it is worth making it.

For instance, supporting other delimited calls from `readr` should be simple, since they all take the same argument structure.   I think you're right about base types, though currently the `read_` side of the code relies on `readr` algorithm to determine the compression type.  Could do this by parsing the file name instead; not sure I want to introduce a dependency on libmagic / `wand` just to detect compression type by magic number...

Other parsers get more tricky -- e.g. I wanted to support `fst` package to considerably speed up the process, but it doesn't seem to support the notion `append`. I haven't looked into `feather` feasibility-wise, but I think its authors say it's not a stable/archival format anyway, which cuts against the spirit of `ark` (and is meanwhile slower than other modern binary formats  like `fst`).  Of course compression is a bigger issue with many of these, particularly those that do not operate over the standard R connection . Ideas here are welcome but I'd be tempted to punt on this until a later date....

A somewhat related challenge that's not directly raised by the reviews is the issue of type stability.  While some of the binary types I just mentioned support this out of the box, types (dates, factors, integer vs numeric etc) can change from a read-in / read-out.  At the moment I feel this is an out-of-scope issue for a first release, but maybe a metadata file should be written with column types when archiving a database?  (This could work within a family like `readr`, harder to support this automatically across different engines -- the heterogeneity of file-based I/O really makes me appreciate the standardization provided by DBI / ODBC -- on the database side the interoperability just works without me having to do anything!

Up against a deadline now but should be able to get this done by July 30.
