### 2018-07-18 cboettig/arkdb#3 @richfitz
Relevant issue for fst: https://github.com/fstpackage/fst/issues/9
