### 2018-07-18 cboettig/arkdb#3 @richfitz
...and https://github.com/fstpackage/fst/issues/24
