### 2018-07-18 076f6ce
Proof-of-concept pluggable i/o
