+++ man/unark.Rd
@@ -4,7 +4,8 @@
 \alias{unark}
 \title{Unarchive a list of compressed tsv files into a database}
 \usage{
-unark(files, db_con, lines = 10000L, ...)
+unark(files, db_con, streamable_table = streamable_readr_tsv(),
+  lines = 10000L, ...)
 }
 \arguments{
 \item{files}{vector of filenames to be read in. Must be \code{tsv}
