+++ NAMESPACE
@@ -8,6 +8,4 @@ importFrom(DBI,dbSendQuery)
 importFrom(DBI,dbWriteTable)
 importFrom(methods,is)
 importFrom(progress,progress_bar)
-importFrom(readr,read_tsv)
-importFrom(readr,write_tsv)
 importFrom(tools,file_ext)
