### 2018-07-18 #224 @richfitz
@cboettig - I've sketched out a proof-of-concept for generalising away from tsv here https://github.com/cboettig/arkdb/pull/5

While doing that I noticed this section:

```r
  compress <- match.arg(compress)
  rs <- DBI::dbSendQuery(db_con, paste("SELECT COUNT(*) FROM", tablename))
  size <- DBI::dbFetch(rs)
  end <- size[[1]][[1]]
```

* You can use `SELECT 1 FROM` for efficiency I believe
* I think this is the source of your result set warnings - why not just `DBI::dbGetQuery(db_con, ...)`?
