### 2018-07-18 #224 @cboettig
@richfitz Thanks much! I'll take a look tonight.

Quick question: do you mean `SELECT COUNT(1) FROM`?  My SQL is terrible, but my understanding is that it's the same to most interpreters (though I believe COUNT doesn't count NA/NULLs, not sure but perhaps I avoid this a bit by selecting all columns?  Remarkable how difficult it is to ask a DB how many rows a table has -- apparently there are other DB specific ways that are only approximate. guess it comes from a world where DBs are constantly being written to and the notion of a number of rows is a more ephemeral concept anyhow)....

re `dbGetQuery` ... because I don't use DBI much and don't know how it works?  is that the same as dbSendQuery + dbFetch (i.e. when no lazy eval of queries is desired, like here).
