### 2018-07-19 #224 @richfitz
Don't listen to me on the `SELECT 1` thing - I should know better to suggest microoptimisations without checking first:

```r
microbenchmark::microbenchmark(
  carl = DBI::dbGetQuery(db$con, "SELECT COUNT(*) FROM flights"),
  rich = DBI::dbGetQuery(db$con, "SELECT COUNT(1) FROM flights"))
## Unit: milliseconds
##  expr      min       lq     mean   median        uq       max neval
##  carl 2.404747 2.536132 2.818021 2.687870  3.002489  4.136245   100
##  rich 8.938237 9.306043 9.851665 9.705794 10.161812 12.921496   100
```

We ran into this problem at work but the tables in question have something on the order of 4 billion rows - so the approximate counts get quite useful...

Re `dbGetQuery` - use the approach as above though otherwise (from `?dbSendQuery`)

> The ‘dbSendQuery()’ method only submits and synchronously executes
> the SQL query to the database engine.  It does _not_ extract any
> records - for that you need to use the ‘dbFetch()’ method, and
> then you must call ‘dbClearResult()’ when you finish fetching the
> records you need. For interactive use, you should almost always
> prefer ‘dbGetQuery()’.

So you're missing the `dbClearResult()`, which would ideally go inside a `tryCatch` or `withCallingHandlers` at which point you're better off just using `dbGetQuery`...
