### 2018-07-19 cede75e
use dbGetQuery
