### 2018-07-19 b68aaf1
use dbGetQuery
