### 2018-07-19 b69cf29
cleaner dialog on import and export
