### 2018-07-21 cboettig/arkdb#6 @cboettig
Rich FitzJohn Onboarding review
### Major comments

- Support for other text formats (Issue #3, See PR #5)

### Minor comments:

- [x] The example in the README produces a significant number of warnings: Warning: Closing open result set, pending rows - these should be silenced if expected, eliminated if not.

- [ ] The progress and general chattiness needs to be tuneable (e.g., with a verbose and a progress argument. Ideally both would be tuneable as progress bars are not very log-friendly but messages are (you can see this play out a bit with the vignette build - the progress output ends up on the parent terminal for me)

- [x] The log output as it is not incredibly informative after it has been run.

*This has been reformatted as Rich recommended*

- [x] How is the default of lines chosen? Is there some heuristic for best performance? My gut instinct is that as it increases in size, the performance will get better, to a point.

*Yeah, I think function overhead means that the fewer chunks you need, the faster it should be.  I chose this default as a number that would keep things to a memory footprint that would work on most laptops today for most tables (obviously wide enough table can break this heuristic) but large enough to be acceptably fast, e.g. on the nyflights data.  Really no great answer here and the user should expect to tune this parameter if performance is an issue.*

- [x]  So rather than testing "known not to support" test for "known to support". You might also include a package-cached capability test (keyed against class) for a SELECT * FROM table WHERE rownum BETWEEN 1 and 2 - if you don't get a syntax error it should be ok to use.

*Brilliant.  I've introduced an internal function to run this test and cache the result in an environmental variable.  If this call runs without error, the windowing version will be used, otherwise it will fall back to the offset version*

### Super minor comments

- [x] use `inherits`, not `is`.

*(oh right, yay, this drops importFrom methods too)

- [ ] Your postgres windowed code is not tested. You might try that on travis?

*(Could also add test with MonetDBLite, which should support the windowing)*

- [ ] Your normalize_con code is not tested for the non-src_dbi path

- [ ] I am confused by the point of the untested early exit in unark_file - when would this be wanted?

- [ ] There are also a few error paths untested (see covr output)

- [ ]  surely one can compute the headers ahead of time (0'th iteration)  (See code suggestions in onboarding thread)

- [ ] I would be probably inclined to split ark_chunk into a "read" and "save" step.
