### 2018-07-21 57ac625
A few fixes suggested by @richfitz so far
* Added a `NEWS.md` file to track changes to the package.
* Log messages improved as suggested by @richfitz
* Better handling of when to use or not use window method for chunking
* Avoid warnings by properly closing connections using `dbGetQuery`
