+++ NEWS.md
@@ -1,0 +1,4 @@
+# arkdb 0.0.0.9000
+
+* Added a `NEWS.md` file to track changes to the package.
+* Log messages improved as suggested by @richfitz
