+++ NAMESPACE
@@ -6,7 +6,6 @@ importFrom(DBI,dbFetch)
 importFrom(DBI,dbListTables)
 importFrom(DBI,dbSendQuery)
 importFrom(DBI,dbWriteTable)
-importFrom(methods,is)
 importFrom(progress,progress_bar)
 importFrom(readr,read_tsv)
 importFrom(readr,write_tsv)
