### 2018-07-21 #224 @expectopatronum
> I'm terrible at adding these, at least until I get bug reports on particular cases. Are there some obvious example scenarios you can think of that should provide better error handling or error messages?

As I come from software engineering I can think of several cases that could be tested (of which some might be an overkill). I would at least test what happens when the file/database connections are not working though. Other tests could be passing wrong parameters (e.g., names of tables that don't exist), ... But as I said, that might be an overkill :)

> This is why `ark` returns `dir` while `unark` returns `db_con`, and is also why `ark` takes the arguments in the order `db_con`, `dir`, while `unark` takes `files` first and `db_con` as the second argument. I could be talked out of this, as I can see how reversing the order is confusing, but I think it as: the first argument is the _input_ (as required by piping), and the second is the destination. `ark` operates on an existing database and outputs files, `unark` operates on files. What do you think?

Yes, that's completely reasonable and fine with me!

Great job!
