### 2018-07-24 63c4e59
implements @krlmlr approach suggested in #7
