### 2018-07-25 47c661d
patch alternate method
