### 2018-07-25 b333313
alternate method with tests
