+++ tests/testthat/test-arkdb.R
@@ -3,10 +3,13 @@ library(dbplyr)
 library(dplyr)
 library(nycflights13)
 library(fs)
+library(RSQLite)
+library(MonetDBLite)
+
+db <- dbplyr::nycflights13_sqlite(".")
 
 testthat::test_that("we can ark and unark a db", {
-  
-  db <- dbplyr::nycflights13_sqlite(".")
+
   dir <- fs::dir_create("nycflights")
   ark(db, dir, lines = 50000)
   
@@ -33,15 +36,16 @@ testthat::test_that("we can ark and unark a db", {
   dim(myflights[[1]])
   ## Classes not preserved, we get read_tsv guesses on class
   
-  unlink("nycflights", TRUE)
+  unlink(dir, TRUE)
   unlink("local.sqlite")
-  unlink("nycflights13.sqlite")
   
 })
 
 
 
 testthat::test_that("we can ark and unark a db in plain text", {
+  
+  dir <- fs::dir_create("nycflights")
   
   db <- dbplyr::nycflights13_sqlite(".")
   dir <- fs::dir_create("nycflights")
@@ -58,12 +62,79 @@ testthat::test_that("we can ark and unark a db in plain text", {
   testthat::expect_equal(dim(flights)[[2]], 19)
   testthat::expect_is(flights, "tbl_dbi")
   
-  unlink("nycflights", TRUE)
+  unlink(dir, TRUE)
   unlink("local.sqlite")
-  unlink("nycflights13.sqlite")
-  
+
 })
 
+
+
+
+
+testthat::test_that("alternate method for ark", {
+  
+  dir <- fs::dir_create("nycflights")
+  ark(db, dir, lines = 50000, use_alternate = TRUE)
+  
+  files <- fs::dir_ls(dir, glob = "*.tsv.bz2")
+  testthat::expect_length(files, 5)
+  
+  myflights <- suppressMessages(
+    readr::read_tsv(fs::path(dir, "flights.tsv.bz2")))
+  testthat::expect_equal(dim(myflights), 
+                         dim(nycflights13::flights))
+  
+  ## unark
+  new_db <- src_sqlite("local.sqlite", create = TRUE)
+  unark(files, new_db, lines = 50000)
+  
+  myflights <- tbl(new_db, "flights")
+  testthat::expect_is(myflights, "tbl_dbi")
+  
+  myflights <- collect(myflights)
+  testthat::expect_equal(dim(myflights), 
+                         dim(nycflights13::flights))
+  
+  
+  ## Classes not preserved, we get read_tsv guesses on class
+ 
+# test alternate DB
+  monet_dir <- fs::dir_create("monet")
+  new_db <- DBI::dbConnect(MonetDBLite::MonetDBLite(), monet_dir)
+
+  ## unark into a new db
+  unark(files, new_db, lines = 50000)
+  
+  flights <- tbl(new_db, "flights")
+  testthat::expect_equal(dim(flights)[[2]], 19)
+  testthat::expect_is(flights, "tbl_dbi")
+  
+  
+  #### ark a monet-db #######
+  
+  ## start clean
+  unlink(dir, TRUE) # ark'd text files
+  dir <- fs::dir_create("nycflights")
+  
+  ark(new_db, dir, lines = 50000, use_alternate = TRUE)
+  
+  ## test
+  files <- fs::dir_ls(dir, glob = "*.tsv.bz2")
+  testthat::expect_length(files, 5)
+  
+  myflights <- suppressMessages(
+    readr::read_tsv(fs::path(dir, "flights.tsv.bz2")))
+  testthat::expect_equal(dim(myflights), 
+                         dim(nycflights13::flights))
+  
+  
[…]
