+++ man/ark.Rd
@@ -5,7 +5,7 @@
 \title{Archive tables from a database as flat files}
 \usage{
 ark(db_con, dir, lines = 10000L, compress = c("bzip2", "gzip", "xz",
-  "none"), tables = list_tables(db_con))
+  "none"), tables = list_tables(db_con), use_alternate = FALSE)
 }
 \arguments{
 \item{db_con}{a database connection}
@@ -20,6 +20,11 @@ no compression.}
 
 \item{tables}{a list of tables from the database that should be
 archived.  By default, will archive all tables.}
+
+\item{use_alternate}{Use a fallback method to query the database. Only useful
+if the \code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}} implementation for your database platform returns the
+full results to the client immediately rather than supporting chunking with \code{n}
+parameter.}
 }
 \value{
 the path to \code{dir} where output files are created (invisibly), for piping.
