+++ tests/testthat/test-arkdb.R
@@ -1,15 +1,16 @@
 library(arkdb)
-library(dbplyr)
-library(dplyr)
-library(nycflights13)
-library(fs)
-library(RSQLite)
-library(MonetDBLite)
+#library(dbplyr)
+#library(dplyr)
+#library(nycflights13)
+#library(fs)
+#library(RSQLite)
+#library(MonetDBLite)
 
-db <- dbplyr::nycflights13_sqlite(".")
 
+testthat::context("basic")
 testthat::test_that("we can ark and unark a db", {
 
+  db <- dbplyr::nycflights13_sqlite(".")
   dir <- fs::dir_create("nycflights")
   ark(db, dir, lines = 50000)
   
@@ -22,13 +23,13 @@ testthat::test_that("we can ark and unark a db", {
                          dim(nycflights13::flights))
   
   ## unark
-  new_db <- src_sqlite("local.sqlite", create = TRUE)
+  new_db <- dplyr::src_sqlite("local.sqlite", create = TRUE)
   unark(files, new_db, lines = 50000)
   
-  myflights <- tbl(new_db, "flights")
+  myflights <- dplyr::tbl(new_db, "flights")
   testthat::expect_is(myflights, "tbl_dbi")
   
-  myflights <- collect(myflights)
+  myflights <- dplyr::collect(myflights)
   testthat::expect_equal(dim(myflights), 
                          dim(nycflights13::flights))
   
@@ -42,9 +43,11 @@ testthat::test_that("we can ark and unark a db", {
 })
 
 
+testthat::context("plain-txt")
 
 testthat::test_that("we can ark and unark a db in plain text", {
   
+  db <- dbplyr::nycflights13_sqlite(".")
   dir <- fs::dir_create("nycflights")
   
   db <- dbplyr::nycflights13_sqlite(".")
@@ -55,10 +58,10 @@ testthat::test_that("we can ark and unark a db in plain text", {
   testthat::expect_length(files, 5)
   
   ## unark
-  new_db <- src_sqlite("local.sqlite", create = TRUE)
+  new_db <- dplyr::src_sqlite("local.sqlite", create = TRUE)
   unark(files, new_db, lines = 50000)
   
-  flights <- tbl(new_db, "flights")
+  flights <- dplyr::tbl(new_db, "flights")
   testthat::expect_equal(dim(flights)[[2]], 19)
   testthat::expect_is(flights, "tbl_dbi")
   
@@ -68,11 +71,11 @@ testthat::test_that("we can ark and unark a db in plain text", {
 })
 
 
-
-
+testthat::context("alternate method")
 
 testthat::test_that("alternate method for ark", {
   
+  db <- dbplyr::nycflights13_sqlite(".")
   dir <- fs::dir_create("nycflights")
   ark(db, dir, lines = 50000, use_alternate = TRUE)
   
@@ -85,13 +88,13 @@ testthat::test_that("alternate method for ark", {
                          dim(nycflights13::flights))
   
   ## unark
-  new_db <- src_sqlite("local.sqlite", create = TRUE)
+  new_db <- dplyr::src_sqlite("local.sqlite", create = TRUE)
   unark(files, new_db, lines = 50000)
   
-  myflights <- tbl(new_db, "flights")
+  myflights <- dplyr::tbl(new_db, "flights")
   testthat::expect_is(myflights, "tbl_dbi")
   
-  myflights <- collect(myflights)
+  myflights <- dplyr::collect(myflights)
   testthat::expect_equal(dim(myflights), 
                          dim(nycflights13::flights))
   
@@ -105,7 +108,7 @@ testthat::test_that("alternate method for ark", {
[…]
