### 2018-07-25 9076573
separate out test to debug appveyor
