### 2018-07-25 7d24c14
fix has_between method
