### 2018-07-25 02ad7cd
Use @krlmlr naming & structure for method
