+++ tests/testthat/test-arkdb.R
@@ -77,7 +77,7 @@ testthat::test_that("alternate method for ark", {
   
   db <- dbplyr::nycflights13_sqlite(".")
   dir <- fs::dir_create("nycflights")
-  ark(db, dir, lines = 50000, use_alternate = TRUE)
+  ark(db, dir, lines = 50000, method = "window")
   
   files <- fs::dir_ls(dir, glob = "*.tsv.bz2")
   testthat::expect_length(files, 5)
@@ -105,8 +105,6 @@ testthat::test_that("alternate method for ark", {
 })
 
 testthat::context("MonetDB")
-
-
 testthat::test_that("try with MonetDB & alternate method", {
   
   ## SETUP, with text files:
@@ -132,7 +130,7 @@ testthat::test_that("try with MonetDB & alternate method", {
   unlink(dir, TRUE) # ark'd text files
   dir <- fs::dir_create("nycflights")
   
-  ark(new_db, dir, lines = 50000L, use_alternate = TRUE)
+  ark(new_db, dir, lines = 50000L, method = "window")
   
   ## test ark results
   files <- fs::dir_ls(dir, glob = "*.tsv.bz2")
@@ -143,7 +141,7 @@ testthat::test_that("try with MonetDB & alternate method", {
   testthat::expect_equal(dim(myflights), 
                          dim(nycflights13::flights))
   
-  
+  DBI::dbDisconnect(new_db)
   unlink(monet_dir, TRUE)
   unlink("local.sqlite")     # unarked db
   unlink(dir, TRUE) # ark'd text files
@@ -151,7 +149,42 @@ testthat::test_that("try with MonetDB & alternate method", {
 })
 
 
-unlink("nycflights13.sqlite") # database
+
+testthat::context("has_windowing")
+testthat::test_that("we can test for windowing", {
+  
+  ## SETUP, with text files:
+  dir <- fs::dir_create("nycflights")
+  data <-  list(airlines = nycflights13::airlines, 
+                airports = nycflights13::airports)
+  tmp <- lapply(names(data), function(x) 
+    readr::write_tsv(data[[x]], fs::path(dir, paste0(x, ".tsv.gz"))))
+  files <- fs::dir_ls(dir, glob = "*.tsv.gz")
+  testthat::expect_length(files, 2)
+  
+  
+  # set up db
+  ## MonetDB LITE has BETWEEN but not ROWNUM
+  
+  # db_dir <- fs::dir_create("monet")
+  # new_db <- DBI::dbConnect(MonetDBLite::MonetDBLite(), db_dir)
+  # unark(files, new_db, lines = 50000)
+  # unlink(db_dir, TRUE)
+  
+  
+  new_db <- dplyr::src_sqlite("local.sqlite", create = TRUE)
+  unark(files, new_db, lines = 50000)
+  
+  testthat::expect_false( has_between(new_db, "airlines") )
+  
+  
+  unlink(dir, TRUE) # ark'd text files
+  unlink("local.sqlite")     # unarked db
+  
+  
+})
+
+
 
 
 ## dbplyr handles this automatically
