### 2018-07-25 61756b8
robust read_chunked return
