### 2018-07-25 bbf3efa
testthat tests need to be self-contained.
