### 2018-07-25 52476fe
simplify test
