### 2018-07-25 480ee96
:fire: :fire: :fire:
no idea why travis and appveyor are each choosing a different test to fail on.  Cannot reproduce.
