### 2018-07-25 2058f68
problems appear to be compression related on linux...
