+++ R/ark.R
@@ -3,6 +3,7 @@
 #' 
 #' @param db_con a database connection
 #' @param dir a directory where we will write the compressed text files output
+#' @param streamable_table interface for serializing/deserializing in chunks
 #' @param lines the number of lines to use in each single chunk
 #' @param compress file compression algorithm. Should be one of "bzip2" (default),
 #' "gzip" (faster write times, a bit less compression), "xz", or "none", for
@@ -43,7 +44,10 @@
 #' 
 #' 
 #' }
-ark <- function(db_con, dir, lines = 10000L, 
+ark <- function(db_con, 
+                dir, 
+                streamable_table = streamable_readr_tsv(),
+                lines = 50000L, 
                 compress = c("bzip2", "gzip", "xz", "none"),
                 tables = list_tables(db_con),
                 method = c("keep-open", "window", "sql-window")){
@@ -52,11 +56,15 @@ ark <- function(db_con, dir, lines = 10000L,
   compress <- match.arg(compress)
   lines <- as.integer(lines)
   
+  stopifnot(inherits(streamable_table, "streamable_table"))
+  
+  ## exclude sqlite's internal tables
   tables <- tables[!grepl("sqlite_", tables)]
   
   lapply(tables, 
          ark_file, 
-         db_con = normalize_con(db_con), 
+         db_con = normalize_con(db_con),
+         streamable_table = streamable_table,
          lines = lines, 
          dir = dir, 
          compress = compress,
@@ -73,6 +81,7 @@ list_tables <- function(db){
 #' @importFrom DBI dbSendQuery dbFetch dbClearResult dbGetQuery
 ark_file <- function(tablename, 
                      db_con, 
+                     streamable_table,
                      lines, 
                      dir, 
                      compress,
@@ -85,7 +94,9 @@ ark_file <- function(tablename,
                 "xz" = ".xz",
                 "none" = "",
                 ".bz2")
-  filename <- file.path(dir, paste0(tablename, ".tsv", ext))
+  
+  dest <- sprintf("%s.%s%s", tablename, streamable_table$extension, ext)
+  filename <- file.path(dir, dest)
   con <- compressed_file(filename, "w")
   on.exit(close(con))
   
@@ -95,10 +106,13 @@ ark_file <- function(tablename,
   t0 <- Sys.time()
  
   switch(method,
-         "keep-open" = keep_open(db_con, lines, p, tablename, con),
-         "window" = window(db_con, lines, compress, p, tablename, con),
-         "sql-window" = sql_window(db_con, lines, compress, p, tablename, con),
-         keep_open(db_con, lines, p, tablename, con)
+          "keep-open" = keep_open(db_con, streamable_table, lines, 
+                                  p, tablename, con),
+             "window" = window(db_con, streamable_table, lines, 
+                               compress, p, tablename, con),
+         "sql-window" = sql_window(db_con, streamable_table, lines, 
+                                   compress, p, tablename, con),
+         keep_open(db_con, streamable_table, lines, p, tablename, con)
   )
   
   message(sprintf("\t...Done! (in %s)", format(Sys.time() - t0)))
@@ -106,11 +120,11 @@ ark_file <- function(tablename,
 }
 
 
-keep_open <- function(db_con, lines, p, tablename, con){
+keep_open <- function(db_con, streamable_table, lines, p, tablename, con){
   ## Create header to avoid duplicate column names
   query <- paste("SELECT * FROM", tablename, "LIMIT 0")
   header <- DBI::dbGetQuery(db_con, query)
-  readr::write_tsv(header, con, append = FALSE)
[…]
