+++ R/streamable_table.R
@@ -1,3 +1,27 @@
+
+#' streamable table
+#'
+#' @param read read function. Arguments should be "`file`" 
+#' (must be able to take a [connection()] object) and "`...`" (for)
+#' additional arguments.
+#' @param write write function. Arguments should be "`data`" (a data.frame),
+#' `file` (must be able to take a [connection()] object), and "`append`" logical,
+#' append to existing file or create new?
+#' @param extension file extension to use (e.g. "tsv", "csv")
+#'
+#' @return a `streamable_table` object (S3)
+#' @export
+#'
+#' @examples
+#' 
+#' streamable_readr_tsv <- function() {
+#'   streamable_table(
+#'     function(file, ...) readr::read_tsv(file, ...),
+#'     function(x, path, append)
+#'       readr::write_tsv(x = x, path = path, append = append),
+#'     "tsv")
+#' }
+#' 
 streamable_table <- function(read, write, extension) {
   stopifnot(is.function(read),
             is.function(write),
@@ -8,6 +32,7 @@ streamable_table <- function(read, write, extension) {
   class(ret) <- "streamable_table"
   ret
 }
+
 
 #' @importFrom readr write_tsv read_tsv
 streamable_readr_tsv <- function() {
@@ -20,13 +45,26 @@ streamable_readr_tsv <- function() {
     }
     
   streamable_table(read_tsv, write_tsv, "tsv")
-
 }
+
+#' @importFrom readr write_csv read_csv
+streamable_readr_csv <- function() {
+  
+  read_csv <- function(file, ...) {
+    readr::read_csv(file, ...)
+  }
+  write_csv <- function(x, path, append = FALSE) {
+    readr::write_csv(x = x, path = path, append = append)
+  }
+  
+  streamable_table(read_csv, write_csv, "csv")
+}
+
 
 # https://www.iana.org/assignments/media-types/text/tab-separated-values
 streamable_base_tsv <- function() {
   read_tsv <- function(file, ...) {
-    utils::read.table(file, 
+    utils::read.table(textConnection(file), 
                       header = TRUE, 
                       sep = "\t", 
                       quote = "",
@@ -34,13 +72,13 @@ streamable_base_tsv <- function() {
                       ...)
   }
   write_tsv <- function(x, path, append) {
-    utils::write.table(file, 
+    utils::write.table(x,
+                       file = path, 
                        append = append, 
                        sep = "\t",
                        quote = FALSE,
                        row.names = FALSE,
-                       col.names = !append,
-                       stringsAsFactors = FALSE)
+                       col.names = !append)
   }
   streamable_table(read_tsv, write_tsv, "tsv")
 }
@@ -48,21 +86,24 @@ streamable_base_tsv <- function() {
 ## FIXME, user might want to set more of these, e.g. header = FALSE...
 streamable_base_csv <- function() {
   read_csv <- function(file, ...) {
-    utils::read.csv(file, 
+    utils::read.table(textConnection(file), 
                       header = TRUE, 
                       sep = ",", 
                       quote = "\"",
                       stringsAsFactors = FALSE,
                       ...)
   }
+  ## write.csv does not permit setting col.names = FALSE, so cannot append
   write_csv <- function(x, path, append) {
-    utils::write.csv(file, 
-                       append = append, 
-                       sep = ",",
-                       quote = "\"",
+      utils::write.table(x,
+                       file = path, 
+                       sep = ",", 
+                       quote = TRUE,
+                       qmethod = "double",
                        row.names = FALSE,
                        col.names = !append,
-                       stringsAsFactors = FALSE)
+                       append = append
+      )
   }
   streamable_table(read_csv, write_csv, "csv")
 }
