+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(ark)
+export(streamable_table)
 export(unark)
 importFrom(DBI,dbClearResult)
 importFrom(DBI,dbFetch)
@@ -9,6 +10,8 @@ importFrom(DBI,dbListTables)
 importFrom(DBI,dbSendQuery)
 importFrom(DBI,dbWriteTable)
 importFrom(progress,progress_bar)
+importFrom(readr,read_csv)
 importFrom(readr,read_tsv)
+importFrom(readr,write_csv)
 importFrom(readr,write_tsv)
 importFrom(tools,file_ext)
