### 2018-07-26 2ae5122
document and export some alternative streamable methods
