+++ R/streamable_table.R
@@ -33,7 +33,12 @@ streamable_table <- function(read, write, extension) {
   ret
 }
 
-
+#' streamable tsv using `readr`
+#' 
+#' @return a `streamable_table` object (S3)
+#' @export
+#' @seealso [readr::read_tsv()], [readr::write_tsv()]
+#' 
 #' @importFrom readr write_tsv read_tsv
 streamable_readr_tsv <- function() {
   
@@ -47,6 +52,12 @@ streamable_readr_tsv <- function() {
   streamable_table(read_tsv, write_tsv, "tsv")
 }
 
+#' streamable csv using `readr`
+#' 
+#' @return a `streamable_table` object (S3)
+#' @export
+#' @seealso [readr::read_csv()], [readr::write_csv()]
+#' 
 #' @importFrom readr write_csv read_csv
 streamable_readr_csv <- function() {
   
@@ -60,8 +71,16 @@ streamable_readr_csv <- function() {
   streamable_table(read_csv, write_csv, "csv")
 }
 
-
-# https://www.iana.org/assignments/media-types/text/tab-separated-values
+#' streamable tsv using base R functions
+#' 
+#' @return a `streamable_table` object (S3)
+#' @export
+#' 
+#' @details
+#' Follows the tab-separate-values standard using [utils::read.table()],
+#' see IANA specification at:
+#' <https://www.iana.org/assignments/media-types/text/tab-separated-values>
+#' 
 streamable_base_tsv <- function() {
   read_tsv <- function(file, ...) {
     utils::read.table(textConnection(file), 
@@ -83,9 +102,19 @@ streamable_base_tsv <- function() {
   streamable_table(read_tsv, write_tsv, "tsv")
 }
 
-## FIXME, user might want to set more of these, e.g. header = FALSE...
+
+
+#' streamable csv using base R functions
+#' 
+#' @return a `streamable_table` object (S3)
+#' @export
+#' 
+#' @details
+#' Follows the comma-separate-values standard using [utils::read.table()]
+#' 
 streamable_base_csv <- function() {
   read_csv <- function(file, ...) {
+    ## Consider case of header = FALSE...
     utils::read.table(textConnection(file), 
                       header = TRUE, 
                       sep = ",", 
@@ -93,7 +122,8 @@ streamable_base_csv <- function() {
                       stringsAsFactors = FALSE,
                       ...)
   }
-  ## write.csv does not permit setting col.names = FALSE, so cannot append
+  ## NOTE: write.csv does not permit setting 
+  ## `col.names = FALSE``, so cannot append
   write_csv <- function(x, path, append) {
       utils::write.table(x,
                        file = path, 
