+++ man/streamable_readr_csv.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/streamable_table.R
+\name{streamable_readr_csv}
+\alias{streamable_readr_csv}
+\title{streamable csv using \code{readr}}
+\usage{
+streamable_readr_csv()
+}
+\value{
+a \code{streamable_table} object (S3)
+}
+\description{
+streamable csv using \code{readr}
+}
+\seealso{
+\code{\link[readr:read_csv]{readr::read_csv()}}, \code{\link[readr:write_csv]{readr::write_csv()}}
+}
