### 2018-07-26 11874df
better docs, better error handling, better testing
