+++ R/streamable_table.R
@@ -5,10 +5,18 @@
 #' (must be able to take a [connection()] object) and "`...`" (for)
 #' additional arguments.
 #' @param write write function. Arguments should be "`data`" (a data.frame),
-#' `file` (must be able to take a [connection()] object), and "`append`" logical,
-#' append to existing file or create new?
+#' `file` (must be able to take a [connection()] object), and "`append`"
+#'  logical, append to existing file or create new?
 #' @param extension file extension to use (e.g. "tsv", "csv")
-#'
+#' @details 
+#' Note several constraints on this design. The write method must be able
+#' to take a generic R `connection` object (which will allow it to handle
+#' the compression methods used, if any), and the read method must be able
+#' to take a `textConnection` object.  `readr` functions handle these cases
+#' out of the box, so the above method is easy to write.  Also note that
+#' the write method must be able to `append`, i.e. it should use a header
+#' if `append=TRUE`, but omit when it is `FALSE`.  See the built-in methods
+#' for more examples.
 #' @return a `streamable_table` object (S3)
 #' @export
 #'
@@ -25,12 +33,21 @@
 streamable_table <- function(read, write, extension) {
   stopifnot(is.function(read),
             is.function(write),
-            is.character(extension), length(extension) == 1L, !is.na(extension))
+            is.character(extension), 
+            length(extension) == 1L, 
+            !is.na(extension))
   ret <- list(read = read,
               write = write,
               extension = extension)
   class(ret) <- "streamable_table"
   ret
+}
+
+
+assert_streamable <- function(x, name = deparse(substitute(x))) {
+  if (!inherits(x, "streamable_table") ) {
+    stop(sprintf("'%s' must be a streamable_table object", name), call. = FALSE)
+  }
 }
 
 #' streamable tsv using `readr`
@@ -39,17 +56,26 @@ streamable_table <- function(read, write, extension) {
 #' @export
 #' @seealso [readr::read_tsv()], [readr::write_tsv()]
 #' 
-#' @importFrom readr write_tsv read_tsv
 streamable_readr_tsv <- function() {
   
-    read_tsv <- function(file, ...) {
-      readr::read_tsv(file, ...)
+  ## Avoids a hard dependency on readr for this courtesy function
+  if (!requireNamespace("readr", quietly = TRUE)) {
+    stop("readr package must be installed to use readr-based methods",
+         call. = FALSE)
+  }
+  read_tsv <- getExportedValue("readr", "read_tsv")
+  write_tsv <- getExportedValue("readr", "write_tsv")
+  
+   
+   ## actual definitions
+    read <- function(file, ...) {
+      read_tsv(file, ...)
     }
-    write_tsv <- function(x, path, append = FALSE) {
-      readr::write_tsv(x = x, path = path, append = append)
+    write <- function(x, path, append = FALSE) {
+      write_tsv(x = x, path = path, append = append)
     }
     
-  streamable_table(read_tsv, write_tsv, "tsv")
+  streamable_table(read, write, "tsv")
 }
 
 #' streamable csv using `readr`
@@ -57,18 +83,25 @@ streamable_readr_tsv <- function() {
 #' @return a `streamable_table` object (S3)
 #' @export
 #' @seealso [readr::read_csv()], [readr::write_csv()]
-#' 
-#' @importFrom readr write_csv read_csv
 streamable_readr_csv <- function() {
   
-  read_csv <- function(file, ...) {
-    readr::read_csv(file, ...)
+  ## Avoids a hard dependency on readr for this courtesy function
+  if (!requireNamespace("readr", quietly = TRUE)) {
[…]
