+++ NAMESPACE
@@ -14,8 +14,6 @@ importFrom(DBI,dbListTables)
 importFrom(DBI,dbSendQuery)
 importFrom(DBI,dbWriteTable)
 importFrom(progress,progress_bar)
-importFrom(readr,read_csv)
-importFrom(readr,read_tsv)
-importFrom(readr,write_csv)
-importFrom(readr,write_tsv)
 importFrom(tools,file_ext)
+importFrom(utils,read.table)
+importFrom(utils,write.table)
