### 2018-07-26 96ae95c
fix error handling, whitespace
