+++ R/unark.R
@@ -122,9 +122,10 @@ unark_file <- function(filename, db_con, streamable_table, lines = 10000L, ...){
 # /blob/4fe82fd29992635b30e637d5412312b0c5e3e38f/R/util.R#L48-L60
 
 assert_files_exist <- function(files){
-  if(length(files) < 1)
-    stop(sprintf("files not found"), call. = FALSE)
-  
+  if(length(files) < 1){
+    if(!file.exists(files))
+      stop(sprintf("files not found"), call. = FALSE)
+  }
   lapply(files, function(f) 
     if(!file.exists(f)) 
       stop(sprintf("'%s' not found", f), call. = FALSE))
