+++ R/streamable_table.R
@@ -177,3 +177,4 @@ streamable_base_csv <- function() {
   }
   streamable_table(read_csv, write_csv, "csv")
 }
+
