### 2018-07-26 #224 @cboettig
@richfitz @expectopatronum @lmullen

Okay, I think I've hit just about all of these now.

@expectopatronum I have added error handling assertions for the argument types at the top of `ark` and `unark`.

@richfitz I've summarized my replies to your issues in https://github.com/cboettig/arkdb/issues/6.  In particular, see https://github.com/cboettig/arkdb/pull/10 which implements your pluggable formats (supporting `csv` and `tsv` and base methods), and https://github.com/cboettig/arkdb/pull/8, which I think largely sidesteps the issue about the different routines trying to window the DB queries (a big thanks to @krlmlr for the pointers there).  The original methods are still supported as an option. Messaging and error handling is also improved following the @richfitz suggestions.

Okay, I think this is a much better package now, though there are no doubt things that could be better still or new issues I've introduced.  Over to you all now; thanks for all the input and looking forward to your feedback.
