### 2018-07-31 295f0bc
reorganize assert files
missing file should be a warning not an error.
