### 2018-08-10 ca4a8ff
Document handling postgres schema, closes #12
