+++ DESCRIPTION
@@ -11,8 +11,8 @@ Authors@R: c(person("Carl", "Boettiger",
                     role = c("aut", "cre", "cph"),
                     comment=c(ORCID = "0000-0002-1642-628X")),
               person("Richard", "FitzJohn", role = "ctb"))
-URL: https://github.com/cboettig/arkdb
-BugReports: https://github.com/cboettig/arkdb/issues
+URL: https://github.com/ropensci/arkdb
+BugReports: https://github.com/ropensci/arkdb/issues
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
