### 2018-08-11 0aa6179
preparing for cran release
