+++ cran-comments.md
@@ -1,0 +1,15 @@
+Dear CRAN maintainers,
+
+We are pleased to submit the initial release of arkdb for CRAN.
+
+
+## Test environments
+* local OS X install, R 3.5.1
+* ubuntu 14.04 (on travis-ci), R 3.5.1
+* win-builder (devel and release)
+
+## R CMD check results
+
+0 errors | 0 warnings | 1 note
+
+* This is a new release.
