+++ README.Rmd
@@ -25,8 +25,8 @@ The goal of arkdb is to provide a convienent way to move data from large compres
 
 ## Links
 
-- A more detailed introduction to package design and use can be found in the package [Vignette](https://cboettig.github.io/arkdb/articles/arkdb_intro.html)
-- [Online versions of package documentation](https://cboettig.github.io/arkdb)
+- A more detailed introduction to package design and use can be found in the package [Vignette](https://ropensci.github.io/arkdb/articles/arkdb_intro.html)
+- [Online versions of package documentation](https://ropensci.github.io/arkdb)
 
 ## Installation
 
