+++ vignettes/arkdb.Rmd
@@ -144,8 +144,8 @@ unark(files, new_db,
 stream <- 
    streamable_table(
      function(file, ...) readr::read_tsv(file, ...),
-     function(x, path, append)
-       readr::write_tsv(x = x, path = path, append = append),
+     function(x, path, omit_header)
+       readr::write_tsv(x = x, path = path, append = omit_header),
      "tsv")
 
 ```
