### 2018-08-13 dbb6449
fix DBI disconnect
