### 2018-08-13 #224 @cboettig
Thanks @lmullen!, I've ticked off the items above and transferred the package over to rOpenSci now.  Badges / CI etc seem to be working fine.  Planning a CRAN release soon.
