+++ tests/testthat/test-errors.R
@@ -4,7 +4,8 @@ testthat::test_that("we can detect errors on types", {
   
   x <- NULL
   testthat::expect_warning(assert_files_exist("not-a-file"), "not found")
-  testthat::expect_warning(assert_files_exist(c("not-a-file", "not")), "not found")
+  testthat::expect_warning(assert_files_exist(
+    c("not-a-file", "not")), "not found")
   testthat::expect_warning(assert_files_exist(x), "no file specified")
   
   testthat::expect_error(assert_dir_exists("not-a-dir"), "not found")
