+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: arkdb
-Version: 0.0.0.9000
+Version: 0.0.1
 Title: Archive and Unarchive Databases Using Flat Files
 Description: Flat text files provide a more robust, compressible,
   and portable way to store tables.  This package provides convenient
