### 2018-08-14 52cdc65
Add zipfile connection function
