### 2018-08-17 8413911
patch overwrite
- Use tmp dir in all examples & vignettes
- Introduce proper handling of interactive and non-overwrite options.
