+++ tests/testthat/test-arkdb.R
@@ -6,25 +6,30 @@ library(arkdb)
 #library(RSQLite)
 #library(MonetDBLite)
 
+# Setup
 tmp <- tempdir()
+db <- dbplyr::nycflights13_sqlite(tmp)
+dir <- file.path(tmp, "nycflights")
+dir.create(dir, showWarnings = FALSE)
+new_db <- dplyr::src_sqlite(fs::path(tmp, "local.sqlite"), create = TRUE)
+
+## Note: later tests will overwrite existing tables and files, throwing warnings
 
 testthat::context("basic")
 testthat::test_that("we can ark and unark a db", {
 
-  db <- dbplyr::nycflights13_sqlite(tmp)
-  dir <- fs::dir_create(fs::path(tmp, "nycflights"))
   ark(db, dir, lines = 50000, overwrite = TRUE)
   
-  files <- fs::dir_ls(dir, glob = "*.tsv.bz2")
+  #files <- fs::dir_ls(dir, glob = "*.tsv.bz2")
+  files <- list.files(dir, pattern = "[.]tsv.bz2$", full.names = TRUE)
   testthat::expect_length(files, 5)
 
   myflights <- suppressMessages(
-    readr::read_tsv(fs::path(dir, "flights.tsv.bz2")))
+    readr::read_tsv(file.path(dir, "flights.tsv.bz2")))
   testthat::expect_equal(dim(myflights), 
                          dim(nycflights13::flights))
   
   ## unark
-  new_db <- dplyr::src_sqlite(fs::path(tmp, "local.sqlite"), create = TRUE)
   unark(files, new_db, lines = 50000, overwrite = TRUE)
   
   myflights <- dplyr::tbl(new_db, "flights")
@@ -34,13 +39,6 @@ testthat::test_that("we can ark and unark a db", {
   testthat::expect_equal(dim(myflights), 
                          dim(nycflights13::flights))
   
- 
-  dim(myflights[[1]])
-  ## Classes not preserved, we get read_tsv guesses on class
-  
-  DBI::dbDisconnect(new_db$con)
-  unlink(dir, TRUE)
-  unlink(fs::path(tmp, "local.sqlite"))
 
 })
 
@@ -48,48 +46,22 @@ testthat::test_that("we can ark and unark a db", {
 testthat::context("plain-txt")
 
 testthat::test_that("we can ark and unark a db in plain text", {
-  db <- dbplyr::nycflights13_sqlite(tmp)
-  dir <- fs::dir_create(fs::path(tmp, "nycflights"))
+  #db <- dbplyr::nycflights13_sqlite(tmp)
+  #dir <- fs::dir_create(fs::path(tmp, "nycflights"))
   
   ark(db, dir, lines = 50000, compress = "none", overwrite = TRUE)
   
-  files <- fs::dir_ls(dir, glob = "*.tsv")
+  #files <- fs::dir_ls(dir, glob = "*.tsv")
+  files <- list.files(dir, pattern = "[.]tsv.bz2$", full.names = TRUE)
   testthat::expect_length(files, 5)
-  
-  ## unark
-  new_db <- dplyr::src_sqlite(fs::path(tmp, "local.sqlite"), create = TRUE)
-  unark(files, new_db, lines = 50000, overwrite = TRUE)
-  
-  flights <- dplyr::tbl(new_db, "flights")
-  testthat::expect_equal(dim(flights)[[2]], 19)
-  testthat::expect_is(flights, "tbl_dbi")
-  
-  DBI::dbDisconnect(new_db$con)
-  unlink(dir, TRUE)
-  unlink(fs::path(tmp, "local.sqlite"))
-})
-
-
-testthat::context("alternate method")
-
-testthat::test_that("alternate method for ark", {
-  
-  db <- dbplyr::nycflights13_sqlite(tmp)
[…]
