### 2018-08-18 07f31ed
cleaning up tests
