### 2018-08-19 96833a6
Patch overwrite method (#16)
