+++ tests/testthat/test-arkdb.R
@@ -6,25 +6,32 @@ library(arkdb)
 #library(RSQLite)
 #library(MonetDBLite)
 
+# Setup
+tmp <- tempdir()
+db <- dbplyr::nycflights13_sqlite(tmp)
+dir <- file.path(tmp, "nycflights")
+dir.create(dir, showWarnings = FALSE)
+new_db <- dplyr::src_sqlite(fs::path(tmp, "local.sqlite"), create = TRUE)
+
+## Note: later tests will overwrite existing tables and files, throwing warnings
 
 testthat::context("basic")
 testthat::test_that("we can ark and unark a db", {
 
-  db <- dbplyr::nycflights13_sqlite(".")
-  dir <- fs::dir_create("nycflights")
-  ark(db, dir, lines = 50000)
+  ark(db, dir, lines = 50000, overwrite = TRUE)
   
-  files <- fs::dir_ls(dir, glob = "*.tsv.bz2")
-  testthat::expect_length(files, 5)
 
-  myflights <- suppressMessages(
-    readr::read_tsv(fs::path(dir, "flights.tsv.bz2")))
+  suppressWarnings(
+    myflights <- readr::read_tsv(file.path(dir, "flights.tsv.bz2"))
+  )
   testthat::expect_equal(dim(myflights), 
                          dim(nycflights13::flights))
   
   ## unark
-  new_db <- dplyr::src_sqlite("local.sqlite", create = TRUE)
-  unark(files, new_db, lines = 50000)
+  #files <- fs::dir_ls(dir, glob = "*.tsv.bz2")
+  files <- list.files(dir, pattern = "[.]tsv.bz2$", full.names = TRUE)
+  testthat::expect_length(files, 5)
+  unark(files, new_db, lines = 50000, overwrite = TRUE)
   
   myflights <- dplyr::tbl(new_db, "flights")
   testthat::expect_is(myflights, "tbl_dbi")
@@ -33,40 +40,39 @@ testthat::test_that("we can ark and unark a db", {
   testthat::expect_equal(dim(myflights), 
                          dim(nycflights13::flights))
   
- 
-  dim(myflights[[1]])
-  ## Classes not preserved, we get read_tsv guesses on class
-  
-  unlink(dir, TRUE)
-  unlink("local.sqlite")
-  
+
 })
 
 
 testthat::context("plain-txt")
 
 testthat::test_that("we can ark and unark a db in plain text", {
+  #db <- dbplyr::nycflights13_sqlite(tmp)
+  #dir <- fs::dir_create(fs::path(tmp, "nycflights"))
   
-  db <- dbplyr::nycflights13_sqlite(".")
-  dir <- fs::dir_create("nycflights")
+  ark(db, dir, lines = 50000, compress = "none", overwrite = TRUE)
   
-  db <- dbplyr::nycflights13_sqlite(".")
-  dir <- fs::dir_create("nycflights")
-  ark(db, dir, lines = 50000, compress = "none")
-  
-  files <- fs::dir_ls(dir, glob = "*.tsv")
+  #files <- fs::dir_ls(dir, glob = "*.tsv")
+  files <- list.files(dir, pattern = "[.]tsv.bz2$", full.names = TRUE)
   testthat::expect_length(files, 5)
+ 
+  suppressWarnings(
+    myflights <- readr::read_tsv(file.path(dir, "flights.tsv.bz2"))
+  )
+  testthat::expect_equal(dim(myflights), 
+                         dim(nycflights13::flights))
   
   ## unark
-  new_db <- dplyr::src_sqlite("local.sqlite", create = TRUE)
-  unark(files, new_db, lines = 50000)
+  #new_db <- dplyr::src_sqlite(fs::path(tmp, "local.sqlite"), create = TRUE)
[…]
