+++ NAMESPACE
@@ -15,5 +15,6 @@ importFrom(DBI,dbSendQuery)
 importFrom(DBI,dbWriteTable)
 importFrom(progress,progress_bar)
 importFrom(tools,file_ext)
+importFrom(utils,askYesNo)
 importFrom(utils,read.table)
 importFrom(utils,write.table)

[…]
