### 2018-08-19 9f3a5e1
rebuild docs for migrated appveyor
