### 2018-08-20 00e7569
tests should check suggested packages are available
*before* running!
