+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: arkdb
-Version: 0.0.1
+Version: 0.0.2
 Title: Archive and Unarchive Databases Using Flat Files
-Description: Flat text files provide a more robust, compressible,
-  and portable way to store tables.  This package provides convenient
+Description: Flat text files provide a robust, compressible, and portable
+  way to store tables from databases.  This package provides convenient
   functions for exporting tables from relational database connections
   into compressed text files and streaming those text files back into
   a database without requiring the whole table to fit in working memory.
