### 2018-09-07 6167e27
backwards compatibility
Remove dep on utils::askYesNo. Closes #17
