+++ cran-comments.md
@@ -1,17 +1,17 @@
 Dear CRAN maintainers,
 
-This release patches the issue in which a unit-test attempted to use a Suggested
-package without first checking that said package was installed.  This also corrects
-a grammatical error in the package DESCRIPTION.  
+This release avoids the use of utils::askYesNo in order to maintain
+greater backwards compatibility with R versions prior to 3.5.0
 
 
 ## Test environments
+
 * local OS X install, R 3.5.1
 * ubuntu 14.04 (on travis-ci), R 3.5.1
 * win-builder (devel and release)
 
 ## R CMD check results
 
-0 errors | 0 warnings | 1 note
+0 errors | 0 warnings | 0 notes
 
 * This is a new release.
