+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: arkdb
-Version: 0.0.2
+Version: 0.0.3
 Title: Archive and Unarchive Databases Using Flat Files
 Description: Flat text files provide a robust, compressible, and portable
   way to store tables from databases.  This package provides convenient
