### 2018-09-07 79600c4
update cran comments
