+++ cran-comments.md
@@ -10,8 +10,3 @@ greater backwards compatibility with R versions prior to 3.5.0
 * ubuntu 14.04 (on travis-ci), R 3.5.1
 * win-builder (devel and release)
 
-## R CMD check results
-
-0 errors | 0 warnings | 0 notes
-
-* This is a new release.
