### 2018-09-20 c8010e6
cleaner import message (no path)
