+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: arkdb
-Version: 0.0.3
+Version: 0.0.3.9000
 Title: Archive and Unarchive Databases Using Flat Files
 Description: Flat text files provide a robust, compressible, and portable
   way to store tables from databases.  This package provides convenient
@@ -18,7 +18,7 @@ Encoding: UTF-8
 LazyData: true
 ByteCompile: true
 VignetteBuilder: knitr
-RoxygenNote: 6.0.1.9000
+RoxygenNote: 6.1.0
 Roxygen: list(markdown = TRUE)
 Imports: 
     DBI,

[…]
