### 2018-09-26 2ca50c2
switch example to `latin1`
Though based on stringi:stri_enc_detect, encoding may actually be ISO-8859-2 instead of ISO-8859-1 (latin1)? Though that causes other parsing errors...
