### 2018-09-26 a9ffd14
all is good with `latin2`.
stringi's guess was correct, we just needed to use R's short name instead of the official encoding name in `options`.
