### 2018-09-26 1ca09cc
adds option to pass encoding directly.
also adds ability for unark to guess csv vs tsv.
