### 2018-09-26 04f353c
tweaks to unark for more robust parsing (#19)
- `unark()` will strip out non-compliant characters in table names by default.

- `unark()` gains the optional argument `tablenames`, allowing the user to
   specify the corresponding table names manually, rather than enforcing
   they correspond with the incoming file names. 
   closes #18

-  `unark()` gains the argument `encoding`, allowing users to directly set
   the encoding of incoming files.  Previously this could only be set by
   setting `options(encoding)`, which will still work as well. See
  `FAO.R` example in `examples` for an illustr
[…]
