+++ man/ark.Rd
@@ -4,9 +4,10 @@
 \alias{ark}
 \title{Archive tables from a database as flat files}
 \usage{
-ark(db_con, dir, streamable_table = streamable_base_tsv(), lines = 50000L,
-  compress = c("bzip2", "gzip", "xz", "none"), tables = list_tables(db_con),
-  method = c("keep-open", "window", "sql-window"), overwrite = "ask")
+ark(db_con, dir, streamable_table = streamable_base_tsv(),
+  lines = 50000L, compress = c("bzip2", "gzip", "xz", "none"),
+  tables = list_tables(db_con), method = c("keep-open", "window",
+  "sql-window"), overwrite = "ask")
 }
 \arguments{
 \item{db_con}{a database connection}
