+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: arkdb
-Version: 0.0.3.9000
+Version: 0.0.4
 Title: Archive and Unarchive Databases Using Flat Files
 Description: Flat text files provide a robust, compressible, and portable
   way to store tables from databases.  This package provides convenient
