### 2018-10-18 #224 @maelle
@lmullen can this issue be closed? 🙏 I'm asking because of my packages registry experiments, where I noticed the review badge of `arkdb` was still orange. 😉
